refact(search): factor out SearchResultsTablePage
